Don't turn on verbose logging based on system.debug
The amount of logging msbuild outputs for diagnostic
verbosity is large and not helpful for debugging a pipeline,
which is why people usually turn on system.debug.
This is a problem because it leads to massive, unexpected
file uploads.
People who want verbose msbuild output can still set the task options.